requirements.txt
# ── Core LLM Framework ──
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-anthropic>=0.3.0    # For Claude API (demo/production)

# ── Embeddings & Vector Store ──
sentence-transformers>=3.0.0
chromadb>=0.5.0

# ── Document Parsing ──
pymupdf>=1.24.0               # PDF reading
unstructured>=0.15.0          # Multi-format document parsing
python-docx>=1.1.0            # DOCX reading

# ── Agent Orchestration (Project 2) ──
langgraph>=0.2.0

# ── Job Search Tools (Project 2) ──
# serpapi                      # Uncomment when you reach Step 6

# ── Frontend ──
streamlit>=1.38.0

# ── Utilities ──
python-dotenv>=1.0.0
tiktoken>=0.7.0               # Token counting
rich>=13.0.0                  # Pretty terminal output
